Data & Domain Layer - Define UserProgression Model & Update Session Model mapper logics

+++ android/app/src/main/java/kr/valor/juggernaut/data/session/mapper/delegate/property/RoutinePropertyMediateDelegate.kt
@@ -1,5 +1,5 @@
 package kr.valor.juggernaut.data.session.mapper.delegate.property
 
 interface RoutinePropertyMediateDelegate {
-    fun mediate(input: Double): Double
+    fun mediate(input: Double): Int
 }
