+++ android/app/src/main/java/kr/valor/juggernaut/domain/session/model/Routine.kt
@@ -1,6 +1,6 @@
 package kr.valor.juggernaut.domain.session.model
 
 data class Routine(
-    val weights: Double,
+    val weights: Int,
     val reps: Int
 )
